backend/requirements.txt
fastapi
uvicorn
sqlmodel[sqlite]
python-jose
pydantic
passlib
bcrypt<4.1
